+++ ros-20190910/alibabacloud_ros20190910/client.py
@@ -343,6 +343,7 @@ class Client(OpenApiClient):
         query['TemplateBody'] = request.template_body
         query['TemplateId'] = request.template_id
         query['TemplateScratchId'] = request.template_scratch_id
+        query['TemplateScratchRegionId'] = request.template_scratch_region_id
         query['TemplateURL'] = request.template_url
         query['TemplateVersion'] = request.template_version
         query['TimeoutInMinutes'] = request.timeout_in_minutes
@@ -389,6 +390,7 @@ class Client(OpenApiClient):
         query['TemplateBody'] = request.template_body
         query['TemplateId'] = request.template_id
         query['TemplateScratchId'] = request.template_scratch_id
+        query['TemplateScratchRegionId'] = request.template_scratch_region_id
         query['TemplateURL'] = request.template_url
         query['TemplateVersion'] = request.template_version
         query['TimeoutInMinutes'] = request.timeout_in_minutes
@@ -2752,6 +2754,7 @@ class Client(OpenApiClient):
         query['TemplateBody'] = request.template_body
         query['TemplateId'] = request.template_id
         query['TemplateScratchId'] = request.template_scratch_id
+        query['TemplateScratchRegionId'] = request.template_scratch_region_id
         query['TemplateURL'] = request.template_url
         query['TemplateVersion'] = request.template_version
         req = open_api_models.OpenApiRequest(
@@ -2786,6 +2789,7 @@ class Client(OpenApiClient):
         query['TemplateBody'] = request.template_body
         query['TemplateId'] = request.template_id
         query['TemplateScratchId'] = request.template_scratch_id
+        query['TemplateScratchRegionId'] = request.template_scratch_region_id
         query['TemplateURL'] = request.template_url
         query['TemplateVersion'] = request.template_version
         req = open_api_models.OpenApiRequest(
@@ -4435,6 +4439,7 @@ class Client(OpenApiClient):
         query['TemplateBody'] = request.template_body
         query['TemplateId'] = request.template_id
         query['TemplateScratchId'] = request.template_scratch_id
+        query['TemplateScratchRegionId'] = request.template_scratch_region_id
         query['TemplateURL'] = request.template_url
         query['TemplateVersion'] = request.template_version
         query['TimeoutInMinutes'] = request.timeout_in_minutes
@@ -4475,6 +4480,7 @@ class Client(OpenApiClient):
         query['TemplateBody'] = request.template_body
         query['TemplateId'] = request.template_id
         query['TemplateScratchId'] = request.template_scratch_id
+        query['TemplateScratchRegionId'] = request.template_scratch_region_id
         query['TemplateURL'] = request.template_url
         query['TemplateVersion'] = request.template_version
         query['TimeoutInMinutes'] = request.timeout_in_minutes